+++ src/file/ini_reader.c
@@ -10,6 +10,7 @@ home page for more info:
 https://github.com/benhoyt/inih
 
 */
+#include "mem/mem_debug.h"
 
 #if defined(_MSC_VER) && !defined(_CRT_SECURE_NO_WARNINGS)
 #define _CRT_SECURE_NO_WARNINGS